+++ paasta_tools/paastak8s/bin/kubernetes_remove_evicted_pods.py
