+++ paasta_tools/paastak8s/bin/paasta_cleanup_stale_nodes.py
