+++ paasta_tools/paastak8s/bin/paasta_secrets_sync.py
