+++ setup.py
@@ -61,9 +61,9 @@ setup(
         "paasta_tools/generate_services_file.py",
         "paasta_tools/generate_services_yaml.py",
         "paasta_tools/get_mesos_leader.py",
-        "paasta_tools/kubernetes/bin/paasta_secrets_sync.py",
-        "paasta_tools/kubernetes/bin/paasta_cleanup_stale_nodes.py",
-        "paasta_tools/kubernetes/bin/kubernetes_remove_evicted_pods.py",
+        "paasta_tools/paastak8s/bin/paasta_secrets_sync.py",
+        "paasta_tools/paastak8s/bin/paasta_cleanup_stale_nodes.py",
+        "paasta_tools/paastak8s/bin/kubernetes_remove_evicted_pods.py",
         "paasta_tools/list_marathon_service_instances.py",
         "paasta_tools/log_task_lifecycle_events.py",
         "paasta_tools/marathon_dashboard.py",
